pyproject.toml
[project]
name = "cohorts"
version = "0.1.0"
description = "The Rebble cohorts API"
readme = "README.md"
requires-python = "~=3.14.0"
dependencies = [
    "boto3>=1.35.0",
    "flask>=3.1.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.3.0",
    "honeycomb-beeline>=3.6.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "ruff>=0.15.10",
]

[project.scripts]
debug_server = "serve_debug:main"
